chore: refactor normalizeScriptPath where it makes more sense
