+++ lib/debugger-wrapper.js
@@ -1,4 +1,3 @@
-const path = require('path');
 const debug = require('debug')('snyk:nodejs:inspector');
 const inspector = require('inspector');
 
@@ -29,7 +28,7 @@ function init() {
 
   session.on('Debugger.scriptParsed', (script) => {
     loadedScripts[script.params.scriptId] = script.params;
-    const scriptPath = normalizeScriptPath(script.params.url);
+    const scriptPath = moduleUtils.normalizeScriptPath(script.params.url);
 
     // dealing only with 3rd party modules.
     if (moduleUtils.isNative(scriptPath)) {
@@ -122,16 +121,6 @@ function handleDebuggerPausedEvent(message) {
   };
 
   transmitter.addEvent({methodEntry});
-}
-
-function normalizeScriptPath(scriptPath) {
-  // Remove file prefix which was added in Node v10.12
-  if (scriptPath.startsWith('file://') && path.sep === '/') {
-    scriptPath = scriptPath.substring('file://'.length);
-  } else if (scriptPath.startsWith('file:///') && path.sep === '\\') {
-    scriptPath = scriptPath.substring('file:///'.length).replace(/\//g, '\\');
-  }
-  return scriptPath;
 }
 
 module.exports = {handlePeriodicTasks, init};
