+++ lib/module-utils.js
@@ -1,5 +1,5 @@
-var path = require('path');
 var fs = require('fs');
+var path = require('path');
 
 function isNative(filePath) {
   return !filePath.includes('node_modules');
@@ -22,4 +22,15 @@ function getModuleInfo(filePath) {
   };
   return moduleInfo;
 }
-module.exports = {isNative, getModuleInfo};
+
+function normalizeScriptPath(scriptPath) {
+  // Remove file prefix which was added in Node v10.12
+  if (scriptPath.startsWith('file://') && path.sep === '/') {
+    scriptPath = scriptPath.substring('file://'.length);
+  } else if (scriptPath.startsWith('file:///') && path.sep === '\\') {
+    scriptPath = scriptPath.substring('file:///'.length).replace(/\//g, '\\');
+  }
+  return scriptPath;
+}
+
+module.exports = {isNative, getModuleInfo, normalizeScriptPath};
